Style first content box
Use flexbox to style the first content box. There are two containers,
each having flex. Left container has information and is aligned
in columns. The right container flexes to justify image to the right.
Both are flex to 1 as the child of the first content box so they
have equal boxes.